refactor(graph): split prefigure/runtime and point controls modules
Extract prefigure runtime, sanitize, and diagcess helpers.

Split GraphControls into shell, point-specific component, state hook, and presentation components while preserving behavior.
