+++ packages/doenetml/src/Viewer/renderers/prefigure.tsx
@@ -1,37 +1,20 @@
 import React, { useEffect, useRef, useState } from "react";
-import DOMPurify from "dompurify";
 import {
-    PREFIGURE_BUILD_ENDPOINT,
-    PREFIGURE_DIAGCESS_SCRIPT_URL,
-    PREFIGURE_INDEX_URL,
-    PREFIGURE_MODULE_URL,
-} from "./utils/prefigureConfig";
-
-const PREFIGURE_BUILD_DEBOUNCE_COLD_MS = 1000;
-const PREFIGURE_BUILD_DEBOUNCE_WARM_MS = 40;
-const DIAGCESS_REINIT_DELAY_MS = 400;
-
-type PrefigureModule = typeof import("@doenet/prefigure");
-
-let prefigureModulePromise: Promise<PrefigureModule> | null = null;
-let prefigureWarmupPromise: Promise<PrefigureModule> | null = null;
-let prefigureReadyModule: PrefigureModule | null = null;
-
-type PrefigureBuildResult = {
-    svg: string;
-    annotationsXml: string;
-};
-
-type PrefigureBuildWinner =
-    | { backend: "service"; data: PrefigureBuildResult }
-    | { backend: "local"; module: PrefigureModule };
-
-type DiagcessApi = {
-    Base: {
-        init: () => void;
-        molMap: Record<string, unknown>;
-    };
-};
+    buildPrefigureDiagram,
+    currentPrefigureDebounceMs,
+    isAbortError,
+    warmupPrefigureInBackground,
+} from "./utils/prefigureRuntime";
+import {
+    hasAnnotationsXml,
+    sanitizeBuildResult,
+} from "./utils/prefigureSanitize";
+import {
+    diagcessApi,
+    ensureDiagcessScriptLoaded,
+    queueDiagcessReinit,
+    removeDiagcessMessages,
+} from "./utils/prefigureDiagcess";
 
 type PrefigureRendererProps = {
     id: string;
@@ -45,535 +28,6 @@ type PrefigureRendererProps = {
     surfaceStyle: React.CSSProperties;
 };
 
-async function importPrefigureFromUrl(url: string): Promise<PrefigureModule> {
-    return import(/* @vite-ignore */ url);
-}
-
-async function getPrefigureModule() {
-    if (!prefigureModulePromise) {
-        prefigureModulePromise = importPrefigureFromUrl(PREFIGURE_MODULE_URL);
-    }
-
-    return prefigureModulePromise;
-}
-
-async function startPrefigureWarmup() {
-    if (!prefigureWarmupPromise) {
-        prefigureWarmupPromise = (async () => {
-            const module = await getPrefigureModule();
-            await module.initPrefigure(PREFIGURE_INDEX_URL || undefined);
-            prefigureReadyModule = module;
-            console.log("[prefigure] WASM runtime ready");
-            return module;
-        })().catch((error) => {
-            // Keep fallback-to-service behavior and allow future retries.
-            prefigureWarmupPromise = null;
-            throw error;
-        });
-    }
-
-    return prefigureWarmupPromise;
-}
-
-function logWarmupFailure(error: unknown) {
-    console.error("[prefigure] warmup failed", error);
-}
-
-function warmupPrefigureInBackground() {
-    startPrefigureWarmup().catch(logWarmupFailure);
-}
-
-function currentPrefigureDebounceMs() {
-    return prefigureReadyModule
-        ? PREFIGURE_BUILD_DEBOUNCE_WARM_MS
-        : PREFIGURE_BUILD_DEBOUNCE_COLD_MS;
-}
-
-async function buildWithPrefigureService(
-    diagramXML: string,
-    signal: AbortSignal,
-): Promise<PrefigureBuildResult> {
-    const response = await fetch(PREFIGURE_BUILD_ENDPOINT, {
-        method: "POST",
-        headers: {
-            "Content-Type": "application/xml",
-        },
-        body: diagramXML,
-        signal,
-    });
-
-    if (!response.ok) {
-        throw new Error(`HTTP Error: ${response.status}`);
-    }
-
-    return await response.json();
-}
-
-/**
- * Creates a DOMException-compatible AbortError for programmatic cancellation.
- * Falls back to a plain Error with `name === "AbortError"` in environments
- * where DOMException is not constructible (e.g. older WebViews).
- */
-function createAbortError(): Error {
-    try {
-        return new DOMException("The operation was aborted.", "AbortError");
-    } catch (_error) {
-        const error = new Error("The operation was aborted.");
-        error.name = "AbortError";
-        return error;
-    }
-}
-
-/**
- * Returns true if `error` is an AbortError regardless of whether it came from
- * a DOMException, from the plain-Error fallback in {@link createAbortError},
- * or from environment-specific abort object shapes.
- */
-function isAbortError(error: unknown): boolean {
-    if (
-        error === null ||
-        (typeof error !== "object" && typeof error !== "function")
-    ) {
-        return false;
-    }
-
-    if (!("name" in error)) {
-        return false;
-    }
-
-    return error.name === "AbortError";
-}
-
-/**
- * Unwraps a multi-error array produced by {@link firstSuccessful} into a
- * single throwable value. Prefers the first non-AbortError so that a real
- * failure surfaces even when one backend was intentionally cancelled.
- */
-function resolveBuildRaceError(error: unknown): unknown {
-    if (!Array.isArray(error)) {
-        return error;
-    }
-
-    const nonAbortError = error.find((candidate) => !isAbortError(candidate));
-    return nonAbortError ?? error[0];
-}
-
-/**
- * Resolves with the value of the first promise that fulfills.
- * Rejects with an array of all errors only when every promise has rejected.
- *
- * Semantically equivalent to `Promise.any()`, but compatible with the
- * `ES2021.String` lib subset used in this package (which lacks `ES2021.Promise`
- * and therefore lacks the `AggregateError` type).
- */
-function firstSuccessful<T>(promises: Array<Promise<T>>): Promise<T> {
-    return new Promise((resolve, reject) => {
-        let remaining = promises.length;
-        const errors: unknown[] = [];
-
-        if (remaining === 0) {
-            reject(new Error("No promises were provided."));
-            return;
-        }
-
-        for (const promise of promises) {
-            promise.then(resolve, (error: unknown) => {
-                errors.push(error);
-                remaining -= 1;
-                if (remaining === 0) {
-                    reject(errors);
-                }
-            });
-        }
-    });
-}
-
-/**
- * Build diagram content via whichever backend is currently available.
- * Prefers local WASM when warm; otherwise uses the build service.
- */
-async function buildPrefigureDiagram(
-    diagramXML: string,
-    signal: AbortSignal,
-): Promise<PrefigureBuildResult> {
-    if (prefigureReadyModule) {
-        return prefigureReadyModule.compilePrefigure(diagramXML, {
-            mode: "svg",
-            indexURL: PREFIGURE_INDEX_URL || undefined,
-        });
-    }
-
-    if (signal.aborted) {
-        throw createAbortError();
-    }
-
-    const serviceAbortController = new AbortController();
-    const abortServiceRequest = () => {
-        serviceAbortController.abort();
-    };
-
-    signal.addEventListener("abort", abortServiceRequest, { once: true });
-
-    // Close the check/listen race: the outer signal may have aborted after the
-    // earlier guard but before this listener was attached.
-    if (signal.aborted) {
-        serviceAbortController.abort();
-    }
-
-    // cleanupOuterAbort removes the abort listener added inside the promise
-    // constructor below so it does not linger on the signal after the race
-    // settles normally.
-    let cleanupOuterAbort: () => void = () => {};
-
-    const outerAbortPromise = new Promise<never>((_resolve, reject) => {
-        const rejectForAbort = () => {
-            reject(createAbortError());
-        };
-
-        if (signal.aborted) {
-            rejectForAbort();
-            return;
-        }
-
-        signal.addEventListener("abort", rejectForAbort, { once: true });
-        cleanupOuterAbort = () => {
-            signal.removeEventListener("abort", rejectForAbort);
-        };
-    });
-
-    try {
-        async function buildServicePromise(): Promise<PrefigureBuildWinner> {
-            const data = await buildWithPrefigureService(
-                diagramXML,
-                serviceAbortController.signal,
-            );
-            return { backend: "service", data };
-        }
-
-        async function buildLocalPromise(): Promise<PrefigureBuildWinner> {
-            try {
-                const module = await startPrefigureWarmup();
-                return { backend: "local" as const, module };
-            } catch (error) {
-                logWarmupFailure(error);
-                throw error;
-            }
-        }
-
-        const serviceBuildPromise = buildServicePromise();
-        const localReadyPromise = buildLocalPromise();
-
-        const winner = await Promise.race([
-            firstSuccessful([serviceBuildPromise, localReadyPromise]),
-            outerAbortPromise,
-        ]);
-
-        if (winner.backend === "service") {
-            return winner.data;
-        }
-
-        // WASM became ready before the service response, so stop waiting on the
-        // slower network fallback and compile locally.
-        serviceAbortController.abort();
-
-        return await Promise.race([
-            winner.module.compilePrefigure(diagramXML, {
-                mode: "svg",
-                indexURL: PREFIGURE_INDEX_URL || undefined,
-            }),
-            outerAbortPromise,
-        ]);
-    } catch (error) {
-        throw resolveBuildRaceError(error);
-    } finally {
-        signal.removeEventListener("abort", abortServiceRequest);
-        cleanupOuterAbort();
-    }
-}
-
-function diagcessApi(): DiagcessApi | undefined {
-    return (window as Window & { diagcess?: DiagcessApi }).diagcess;
-}
-
-const DIAGCESS_SCRIPT_MARKER_ATTR = "data-doenet-diagcess-script";
-const DIAGCESS_SCRIPT_LOADED_ATTR = "data-doenet-diagcess-loaded";
-let diagcessScriptLoadPromise: Promise<void> | null = null;
-
-function resolveScriptUrl(url: string): string {
-    return new URL(url, window.location.href).href;
-}
-
-function findDiagcessScript(): HTMLScriptElement | null {
-    const expectedUrl = resolveScriptUrl(PREFIGURE_DIAGCESS_SCRIPT_URL);
-    const scripts = Array.from(document.getElementsByTagName("script"));
-
-    for (const script of scripts) {
-        const matchesMarker =
-            script.getAttribute(DIAGCESS_SCRIPT_MARKER_ATTR) === "true";
-        const matchesUrl = script.src === expectedUrl;
-
-        if (matchesMarker || matchesUrl) {
-            return script;
-        }
-    }
-
-    return null;
-}
-
-function ensureDiagcessScriptLoaded(): Promise<void> {
-    if (diagcessApi()) {
-        return Promise.resolve();
-    }
-
-    if (diagcessScriptLoadPromise) {
-        return diagcessScriptLoadPromise;
-    }
-
-    diagcessScriptLoadPromise = new Promise<void>((resolve, reject) => {
-        const resolveLoaded = () => {
-            resolve();
-        };
-
-        const rejectLoad = () => {
-            diagcessScriptLoadPromise = null;
-            reject(new Error("Failed to load diagcess script."));
-        };
-
-        const existingScript = findDiagcessScript();
-        if (existingScript) {
-            if (
-                diagcessApi() ||
-                existingScript.getAttribute(DIAGCESS_SCRIPT_LOADED_ATTR) ===
-                    "true"
-            ) {
-                resolveLoaded();
-                return;
-            }
-
-            existingScript.addEventListener("load", resolveLoaded, {
-                once: true,
-            });
-            existingScript.addEventListener("error", rejectLoad, {
-                once: true,
-            });
-            return;
-        }
-
-        const script = document.createElement("script");
-        script.src = PREFIGURE_DIAGCESS_SCRIPT_URL;
-        script.type = "text/javascript";
-        script.async = true;
-        script.setAttribute(DIAGCESS_SCRIPT_MARKER_ATTR, "true");
-
-        script.addEventListener(
-            "load",
-            () => {
-                script.setAttribute(DIAGCESS_SCRIPT_LOADED_ATTR, "true");
-                resolveLoaded();
-            },
-            { once: true },
-        );
-
-        script.addEventListener("error", rejectLoad, { once: true });
-        document.head.appendChild(script);
-    });
-
-    return diagcessScriptLoadPromise;
-}
-
-function normalizeSerializedMarkup(value: unknown): string {
-    if (typeof value !== "string") {
-        return "";
-    }
-
-    let text = value;
-
-    for (let ind = 0; ind < 2; ind++) {
-        const trimmed = text.trim();
-
-        if (!(trimmed.startsWith('"') && trimmed.endsWith('"'))) {
-            break;
-        }
-
-        try {
-            const parsed = JSON.parse(trimmed);
-            if (typeof parsed === "string") {
-                text = parsed;
-            } else {
-                break;
-            }
-        } catch (_e) {
-            break;
-        }
-    }
-
-    return text;
-}
-
-function hasAnnotationsXml(value: string): boolean {
-    const trimmed = value.trim();
-    if (!trimmed) {
-        return false;
-    }
-
-    // The annotations payload uses <diagram> as its root element.
-    return /<diagram\b/i.test(trimmed);
-}
-
-function removeDiagcessMessages(container: HTMLElement | null): void {
-    if (!container) {
-        return;
-    }
-
-    for (const child of Array.from(container.children)) {
-        if (
-            child instanceof HTMLParagraphElement &&
-            child.classList.contains("cacc-message")
-        ) {
-            child.remove();
-        }
-    }
-}
-
-const FORBIDDEN_MARKUP_TAGS = new Set([
-    "script",
-    "foreignobject",
-    "iframe",
-    "object",
-    "embed",
-    "link",
-    "meta",
-    "base",
-]);
-
-const URL_ATTRIBUTE_NAMES = new Set([
-    "href",
-    "xlink:href",
-    "src",
-    "data",
-    "action",
-    "formaction",
-    "poster",
-]);
-
-function isUnsafeUrl(value: string): boolean {
-    const normalized = value.trim().replace(/\s+/g, "").toLowerCase();
-    return (
-        normalized.startsWith("javascript:") ||
-        normalized.startsWith("vbscript:") ||
-        normalized.startsWith("data:") ||
-        normalized.startsWith("file:") ||
-        normalized.startsWith("//")
-    );
-}
-
-function isLocalSvgFragmentReference(value: string): boolean {
-    return value.trim().startsWith("#");
-}
-
-function sanitizeXmlMarkup({
-    markup,
-    mimeType,
-    allowedRootNames,
-}: {
-    markup: string;
-    mimeType: DOMParserSupportedType;
-    allowedRootNames: Set<string>;
-}): string {
-    const parser = new DOMParser();
-    const doc = parser.parseFromString(markup, mimeType);
-
-    if (doc.getElementsByTagName("parsererror").length > 0) {
-        return "";
-    }
-
-    const root = doc.documentElement;
-    const rootName = root?.tagName?.toLowerCase?.();
-    if (!rootName || !allowedRootNames.has(rootName)) {
-        return "";
-    }
-
-    const elements = [root, ...Array.from(root.getElementsByTagName("*"))];
-    for (const element of elements) {
-        const tagName = (element.localName || element.tagName).toLowerCase();
-
-        if (FORBIDDEN_MARKUP_TAGS.has(tagName)) {
-            element.remove();
-            continue;
-        }
-
-        const attributes = Array.from(element.attributes);
-        for (const attribute of attributes) {
-            const name = attribute.name.toLowerCase();
-            const localName = (
-                attribute.localName || attribute.name
-            ).toLowerCase();
-            const value = attribute.value.trim();
-            const isEventHandler = name.startsWith("on");
-            const isUseHrefAttr = tagName === "use" && localName === "href";
-            const hasUnsafeUseHref =
-                isUseHrefAttr && !isLocalSvgFragmentReference(value);
-            const isScriptUrl = isUnsafeUrl(value);
-            const hasUnsafeUrl =
-                (URL_ATTRIBUTE_NAMES.has(name) ||
-                    URL_ATTRIBUTE_NAMES.has(localName)) &&
-                isScriptUrl;
-            const hasStyleAttr = name === "style";
-            if (
-                isEventHandler ||
-                hasUnsafeUseHref ||
-                hasUnsafeUrl ||
-                hasStyleAttr
-            ) {
-                element.removeAttribute(attribute.name);
-            }
-        }
-    }
-
-    return new XMLSerializer().serializeToString(root);
-}
-
-/**
- * Apply a successful build result to component state after sanitization.
- */
-function applyBuildResultToState({
-    data,
-    setSvgMarkup,
-    setSvgMessage,
-    setAnnotationsXml,
-}: {
-    data: PrefigureBuildResult;
-    setSvgMarkup: React.Dispatch<React.SetStateAction<string>>;
-    setSvgMessage: React.Dispatch<React.SetStateAction<string>>;
-    setAnnotationsXml: React.Dispatch<React.SetStateAction<string>>;
-}) {
-    const svg = normalizeSerializedMarkup(data.svg);
-    if (svg) {
-        const sanitizedSvg = sanitizeSvgMarkup(svg);
-        const responsiveSvg = sanitizedSvg
-            ? normalizeSvgViewport(sanitizedSvg)
-            : "";
-
-        if (responsiveSvg) {
-            setSvgMarkup(responsiveSvg);
-            setSvgMessage("");
-        } else {
-            setSvgMarkup("");
-            setSvgMessage("Error: Invalid or unsafe SVG in build response.");
-        }
-    } else {
-        setSvgMarkup("");
-        setSvgMessage("Error: No SVG found in response.");
-    }
-
-    const annotationsMarkup = normalizeSerializedMarkup(data.annotationsXml);
-    if (annotationsMarkup) {
-        setAnnotationsXml(sanitizeAnnotationsMarkup(annotationsMarkup));
-    } else {
-        setAnnotationsXml("");
-    }
-}
-
 function setBuildErrorMessage(
     error: unknown,
     setSvgMarkup: React.Dispatch<React.SetStateAction<string>>,
@@ -586,95 +40,6 @@ function setBuildErrorMessage(
     setSvgMessage(
         `Error: ${errorMessage}. Check the Console (F12) for CORS details if this failed immediately.`,
     );
-}
-
-function sanitizeSvgMarkup(markup: string): string {
-    const purified = DOMPurify.sanitize(markup, {
-        USE_PROFILES: { svg: true, svgFilters: true },
-        ADD_TAGS: ["use"],
-        FORBID_TAGS: ["foreignObject", "iframe", "object", "embed"],
-        FORBID_ATTR: ["style"],
-        ALLOW_UNKNOWN_PROTOCOLS: false,
-        RETURN_TRUSTED_TYPE: false,
-    });
-
-    if (typeof purified !== "string") {
-        return "";
-    }
-
-    return sanitizeXmlMarkup({
-        markup: purified,
-        mimeType: "image/svg+xml",
-        allowedRootNames: new Set(["svg"]),
-    });
-}
-
-function parseSvgLength(value: string | null): number | null {
-    if (!value) {
-        return null;
-    }
-
-    const parsed = Number.parseFloat(value);
-    return Number.isFinite(parsed) && parsed > 0 ? parsed : null;
-}
-
-/**
- * Convert a built PreFigure SVG into a container-responsive SVG.
- *
- * PreFigure output can arrive with fixed width/height attributes, which makes
- * the browser lay it out at a rigid pixel size. In narrower graph containers
- * that can truncate the visible drawing even though the underlying SVG content
- * is complete. By ensuring a viewBox exists and then switching the root SVG to
- * width/height 100%, the browser scales the existing viewport to fit the graph
- * frame instead of clipping it. If no usable viewBox exists and one cannot be
- * derived from numeric width/height attributes, preserve the original sizing
- * so this normalization step does not change SVG layout semantics.
- */
-function normalizeSvgViewport(markup: string): string {
-    const parser = new DOMParser();
-    const doc = parser.parseFromString(markup, "image/svg+xml");
-    if (doc.getElementsByTagName("parsererror").length > 0) {
-        return "";
-    }
-
-    const root = doc.documentElement;
-    if (!root || root.tagName.toLowerCase() !== "svg") {
-        return "";
-    }
-
-    let hasUsableViewBox = Boolean(root.getAttribute("viewBox"));
-
-    if (!hasUsableViewBox) {
-        const width = parseSvgLength(root.getAttribute("width"));
-        const height = parseSvgLength(root.getAttribute("height"));
-
-        if (width !== null && height !== null) {
-            // Preserve the original drawing coordinates before replacing the
-            // fixed outer size with responsive sizing.
-            root.setAttribute("viewBox", `0 0 ${width} ${height}`);
-            hasUsableViewBox = true;
-        }
-    }
-
-    if (hasUsableViewBox) {
-        root.setAttribute("width", "100%");
-        root.setAttribute("height", "100%");
-        if (!root.getAttribute("preserveAspectRatio")) {
-            // Keep the full diagram visible when the container aspect ratio differs
-            // from the original SVG dimensions.
-            root.setAttribute("preserveAspectRatio", "xMidYMid meet");
-        }
-    }
-
-    return new XMLSerializer().serializeToString(root);
-}
-
-function sanitizeAnnotationsMarkup(markup: string): string {
-    return sanitizeXmlMarkup({
-        markup,
-        mimeType: "application/xml",
-        allowedRootNames: new Set(["diagram"]),
-    });
 }
 
 export default React.memo(function Prefigure({
@@ -773,7 +138,9 @@ export default React.memo(function Prefigure({
             const requestSequence = ++requestSequenceRef.current;
             const abortController = new AbortController();
             fetchAbortControllerRef.current = abortController;
-            const isWarmMode = Boolean(prefigureReadyModule);
+            // currentPrefigureDebounceMs() is sourced from runtime readiness;
+            // this keeps warm/cold UI behavior aligned with runtime selection.
+            const isWarmMode = currentPrefigureDebounceMs() < 1000;
 
             // In warm mode, keep the previous render on screen to avoid flash while
             // the fast local compile is in progress.
@@ -791,12 +158,10 @@ export default React.memo(function Prefigure({
                     return;
                 }
 
-                applyBuildResultToState({
-                    data,
-                    setSvgMarkup,
-                    setSvgMessage,
-                    setAnnotationsXml,
-                });
+                const sanitized = sanitizeBuildResult(data);
+                setSvgMarkup(sanitized.svgMarkup);
+                setSvgMessage(sanitized.svgMessage);
+                setAnnotationsXml(sanitized.annotationsXml);
             } catch (error) {
                 if (isAbortError(error)) {
                     return;
@@ -859,19 +224,10 @@ export default React.memo(function Prefigure({
             prefigureContainer
         ) {
             removeDiagcessMessages(prefigureContainer);
-
-            // diagcess mutates molMap during init, so clear any stale
-            // entries before re-running it against newly inserted markup.
-            diagcess.Base.molMap = {};
-            if (diagcessTimerRef.current) {
-                clearTimeout(diagcessTimerRef.current);
-            }
-            // Wait briefly for the sanitized SVG/annotations XML markup to be present
-            // in the live DOM before diagcess scans and annotates it.
-            diagcessTimerRef.current = setTimeout(() => {
-                diagcessTimerRef.current = null;
-                diagcess.Base.init();
-            }, DIAGCESS_REINIT_DELAY_MS);
+            diagcessTimerRef.current = queueDiagcessReinit({
+                diagcess,
+                priorTimer: diagcessTimerRef.current,
+            });
         }
 
         return () => {
